+++ application/message/models.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from typing import TYPE_CHECKING
 
-from sqlalchemy import Integer, ForeignKey, String, TIMESTAMP, Boolean
+from sqlalchemy import Integer, ForeignKey, String, TIMESTAMP, Boolean, BigInteger, BIGINT
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from database.db import Base
@@ -14,7 +14,7 @@ class Message(Base):
     __tablename__ = 'message'
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True)
-    user_from_id: Mapped[int] = mapped_column(Integer, ForeignKey('user.id'))
+    user_from_id: Mapped[BigInteger] = mapped_column(BIGINT, ForeignKey('user.id'))
     personal_chat_id: Mapped[int] = mapped_column(Integer, ForeignKey('personal_chat.id'))
     text: Mapped[str] = mapped_column(String)
     is_read: Mapped[bool] = mapped_column(Boolean, default=False)
